+++ internal/rbac/types_test.go
@@ -14,8 +14,8 @@ func TestStudyRoleMarshal(t *testing.T) {
 		Name:    "owner",
 	}
 	assert.Equal(t, studyRole, makeStudyOwnerRole(id))
-	assert.Equal(t, "study_618880a8-bb0e-413d-9c02-689d86905466:owner", string(studyRole.RoleName()))
-	assert.Equal(t, studyRole, mustMakeStudyRole(RoleName("study_618880a8-bb0e-413d-9c02-689d86905466:owner")))
+	assert.Equal(t, "study_618880a8-bb0e-413d-9c02-689d86905466_owner", string(studyRole.RoleName()))
+	assert.Equal(t, studyRole, mustMakeStudyRole(RoleName("study_618880a8-bb0e-413d-9c02-689d86905466_owner")))
 }
 
 func TestIsStudyRole(t *testing.T) {
